% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{betas}
\alias{betas}
\title{Compute all the Beta (Internal function).}
\usage{
betas(x, r)
}
\arguments{
\item{x}{numeric vector (ie. sample data).}

\item{r}{the of Betas terms.}
}
\value{
a numeric vector of length r (Betas).
}
\description{
Vectorize internal.b
}

