% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bterm}
\alias{bterm}
\title{Compute a single term of the Betas (Internal function).}
\usage{
bterm(x, k)
}
\arguments{
\item{x}{sample (numeric vector).}

\item{k}{the k-th Beta term.}
}
\value{
a numeric vector of length 1 (the k-th Beta term).
}
\description{
Compute a single term of the Betas (Internal function).
}

