% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{kurt}
\alias{kurt}
\title{A traditional moments based kurtosis estimator.}
\usage{
kurt(x)
}
\arguments{
\item{x}{sample (numeric vector).}
}
\value{
the (traditional) moments skewness measure.
}
\description{
A traditional moments based kurtosis estimator.
}

