% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{pstar}
\alias{pstar}
\title{Compute the coefficients of the Betas (Internal function).}
\usage{
pstar(r)
}
\arguments{
\item{r}{the number of coefficients to be calculated (a whole number).}
}
\value{
a vector of coefficients of length r.
}
\description{
Compute the coefficients of the Betas (Internal function).
}

