% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{shape}
\alias{shape}
\title{Sample L-moments ratios.}
\usage{
shape(x, nlmom = 4, ratios = TRUE)
}
\arguments{
\item{x}{Data (vector / matrix / data.frame).}

\item{nlmom}{The number of l-moments (ratios) to compute (Default = 4).}

\item{ratios}{Compute sample l-moments ratios ? (logical, Default = TRUE).}
}
\value{
A list holding sample l-ratios, sample l-moments, and weights.
}
\description{
This function computes the sample l-moments ratios for a given
dataset (vector / matrix / data.frame).
}

