% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{singlelmom}
\alias{singlelmom}
\title{Compute the r-th l-statistic from a given sample (Internal function).}
\usage{
singlelmom(smpl, r)
}
\arguments{
\item{smpl}{sample (numeric vector).}

\item{r}{the r-th sample l-statistic (a single whole number).}
}
\value{
a numeric vector of length 1 (the r-th l-statistic).
}
\description{
Compute the r-th l-statistic from a given sample (Internal function).
}

