% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{skew}
\alias{skew}
\title{A traditional moments based skewness estimator.}
\usage{
skew(x)
}
\arguments{
\item{x}{sample (numeric vector).}
}
\value{
the (traditional) moments skewness measure.
}
\description{
A traditional moments based skewness estimator.
}

