% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{weight.factor}
\alias{weight.factor}
\title{Weight Coefficient for Sample L-moment. (Internal)}
\usage{
weight.factor(r, j, n)
}
\arguments{
\item{r}{The r-th sample l-moment.}

\item{j}{The j-th term.}

\item{n}{The sample size.}
}
\value{
The j-th weight factor of the r-th sample l-moment.
}
\description{
Weight Coefficient for Sample L-moment. (Internal)
}

