% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{weightFactors}
\alias{weightFactors}
\title{Weight Coefficients for Sample L-moment (Internal).}
\usage{
weightFactors(r, n)
}
\arguments{
\item{r}{The r-th sample l-moment.}

\item{n}{The sample size.}
}
\value{
A vector of weight factors.
}
\description{
Weight Coefficients for Sample L-moment (Internal).
}

