% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{weightMatrix}
\alias{weightMatrix}
\title{Weight Coefficients for Sample L-moments.}
\usage{
weightMatrix(nlmom, n)
}
\arguments{
\item{nlmom}{The number of sample l-moments.}

\item{n}{The sample size.}
}
\value{
A matrix of weight factors.
}
\description{
Weight Coefficients for Sample L-moments.
}

