% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{shape}
\alias{shape}
\title{Sample L-moments ratios.}
\usage{
shape(x, ratios = TRUE, n = 4)
}
\arguments{
\item{x}{sample (vector / matrix / data.frame).}

\item{ratios}{Compute l-moments ratios instead of the raw l-moments? (logical, default = TRUE).}

\item{n}{the number of l-moments (ratios) to compute. (a whole number, default = 4).}
}
\value{
a numeric vector of length n / matrix with 4 columns containing
sample l-moments / l-moments ratios.
}
\description{
This function computes the l-moments ratios for a given
dataset (vector / matrix / data.frame).
}

